requirements.txt
requests==2.25.0
loguru==0.5.3
pyserial==3.5
